requirements.txt
# Core simulation dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
networkx>=2.6.0
simpy>=4.0.0

# Machine Learning and RL
gymnasium>=0.29.0
stable-baselines3>=2.0.0
torch>=1.9.0
scikit-learn>=1.0.0

# Data handling and visualization
plotly>=5.0.0
folium>=0.12.0
geopandas>=0.10.0
shapely>=1.8.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
python-dotenv>=0.19.0
click>=8.0.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Development
black>=21.0.0
flake8>=4.0.0
mypy>=0.910
